Moved modal components into own folder
